autodetect pkgconf-2.4.3.exe on windows/not everyone uses distro-config
without using the distro-config that sets PKG_CONFIG, configure won't
pick the windows version of pkgconf and the harfbuzz build then fails
beause the graphite and icu .pc files cannot be found.

Add an autodetection similar to the one for MESON as a convenience for
building on windows. With that it is enough to have it in PATH or
installed into $LODE_HOME/opt/bin
Can always be overridden by explicitly adding PKG_CONFIG=/path/to/file
into autogen.input

Change-Id: Ia8ea0dffd34e7a97fb28c99ac3f6b561c5dab193
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184221
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Jenkins