sw a11y: Drop pThis param to SwAccessibleTableData_Impl::GetCell etc.
Don't pass `this` around as `pThis` param to
SwAccessibleTableData_Impl::GetCell and
SwAccessibleTableData_Impl::CheckRowAndCol only
to be able to pass it in the
lang::IndexOutOfBoundsException.

Drop the param instead.

(If passing `this` to the exception were very relevant,
the methods could also be made non-const instead, but
for now, I see no real reason for that.)

Change-Id: I4057fbdd813af3f0492ff7f3e64cd46b40a03d30
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184296
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins