fix harfbuzz build with --enable-python=fully-internal on linux
if executable python scripts are run as build commands, they can pick up
system python instead of the LO-internal python and that then can cause
incompatibilities, ranging from just being too old and failing with
stuff like
AttributeError: 'str' object has no attribute 'isascii'
or attempting to load incompatible modules
[…]
    from . import _compiler, _parser
  File "/home/michi/development/git/libreoffice/instdir/program/python-core-3.11.12/lib/re/_compiler.py", line 18, in <module>
    assert _sre.MAGIC == MAGIC, "SRE module mismatch"
AssertionError: SRE module mismatch

So remove the executable bit to make sure the scripts are run with the
python specified for the build, either internal or the system one.

Also undo the workaround added in 736998ccef0bacdd2bbf038c98dacfbe654f1a4d
(Only run meson from python when it apparently needs to be)
While this can help running meson itself if the distro has different
python versions but cannot be run with the default python3 / the python
version used to run meson is likely newer and then also avoids most of
the issues above, it still is bad to mix build-tool python and the
commands executed by the buildtool.

For the cornercase of meson not being able to be executed by the python
used for the build, the solution is to either specify
PYTHON=/hatever/python/system/meson/uses (for the non-fully-internal
case) and/or using a release tarball of meson and pointing
MESON=/path/to/meson.py in autogen.input (can be used with both system
and internal python)

Change-Id: Id8e3ea88e0caa4afa872a62207b081c072496903
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184271
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>