sw a11y: Use range-based for in SwAccessibleMap
In SwAccessibleShapeMap_Impl, implement `const_iterator begin() const`
and `const_iterator end() const` instead of the equivalent
`const_iterator cbegin() const` and `const_iterator cend() const`,
as the former are needed for range-based for loops.

Drop the non-const `iterator end()` as it is not needed anymore.

Convert loops to use range-based for, which removes the hassle
to make sure the iterator is manually incremented in all control
paths.

Change-Id: Iefbd11a5e0ae94252771c4c0127d11b36abf8c14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184310
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>