tdf#139633 Draw Navigator rename pages and objects take two
Recent review of the patch done to provide in-place renaming in the
Writer Navigator (commit 492c94abe05e5ac213475cdd85f158b7a462f824)
inspired me to review what had been done for this for the Draw
Navigator. This patch revamps the original work and makes the context
menu item Rename consistent with the behavior of the Writer Navigator
context menu item Rename. It starts in-place editing instead of opening
the Name dialog.

Known issues:

* Gtk3 in-place editing doesn't notify when editing is canceled by
Escape. This leaves the SdPageObjsTLV editing flag true which causes the
Enter key to start entry editing instead of going to the object in the
document.

* For the Navigator in the sidebar, sal in-place editing doesn't revert
to the original text when Escape is pressed. The sidebar FocusManager
handling of Escape for focus in a panel content location is the culprit.

Change-Id: If51f962daf8f9f10a738fa8483fc4626a037bc6a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184321
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>