Fix the ESC key behavior in the sidebar sdNavigator
This patch allows the sidebar FocusManager to handle the ESC key when
the keyboard focus is in the Navigator sidebar panel. I could not make
program execution enter the Navigator window KeyInputHdl when dragging
an object and pressing escape. The test for in-drag is dropped by this
patch. Possibly pre-welded it was needed. Also, I am unsure of the use
of catching ESCAPE in the Navigator to end a slide show. The patch adds
a test to determine if a slide show is running before ending it and
consuming the escape key so no furthing key-input processing is done.

Change-Id: Ib8cc465dd11988ac95877782e7a54b1e3fdfb2a4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184344
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>