tdf#130857 qt weld: Introduce QtInstanceEntryTreeView
This is the weld::EntryTreeView implementation
using native Qt widgets.

A weld::EntryTreeView is a combobox-like widget,
but consisting of two widgets: a weld::Entry and
a weld::TreeView.

In QtInstanceEntryTreeView, implement logic similar
to the GTK (GtkInstanceEntryTreeView) and VCL
(SalInstanceEntryTreeView) implementations:

* Auto-completion based on the tree view entries
  when typing in the entry/edit
* automatically selecting the corresponding tree
  view entry when the entry/edit text matches one
* forward (page) up/down key events on the entry/edit
  to the tree view, so this can be used to navigate
  through the entries

This worked as expected in a quick test of Math's font
dialog that can be triggered as follows:

* "Format" -> "Fonts"
* click the "Modify" menu buttton and select any entry

Support for that dialog will be declared in QtInstanceBuilder
in a separate commit.

Change-Id: Ide3063e27dd36e87539b9eaf7165dc073669323f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184362
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>