tdf#130857 svx: Use container size for search bar item windows
Set the CheckButtonItemWindow and FindTextFieldControl
size to the size of the containers they contain
(see corresponding .ui files svx/uiconfig/ui/checkbuttonbox.ui
and svx/uiconfig/ui/findbox.ui), not only to the
size of the checkbox/combobox inside of those containers.

While the size is the same for gen and gtk3 on Linux, it
isn't in a WIP branch of mine experimenting to support
using native Qt widgets for those when using the qt6 VCL
plugin, resulting in the checkbox/combobox not being shown
completely.

Change-Id: I5ae47adcf21f787d69e130eb2e45737a32676591
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184374
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins