tdf#130857 qt weld: Don't unconditionally make dialogs modal
Despite QDialog::modal [1] defaulting to false like GtkWindow::modal [2]
and QtBuilder::setDialogProperties correctly evaluating the "modal"
property in .ui files, even non-modal dialogs were run as modal
dialogs when using SAL_VCL_QT_USE_WELDED_WIDGETS=1.
(Seen with Writer's "Insert" -> "Fields" -> "More Fields..."
dialog in a WIP branch.)

This is because QDialog::open was used so far to open the dialog,
for which the doc [3] explains:

> Shows the dialog as a window modal dialog, returning immediately.

Use QDialog::show/QWidget::show instead to show the dialog, which
does unconditionally run the dialog as modal.

This also makes the "Close" button work for that dialog.
Previously, it wouldn't close the dialog because

    bAllow = !pDialog->get_visible() || !pDialog->get_modal();

in SfxChildWindow::QueryClose would evaluate to false
because the dialog was both, visible and modal.

[1] https://doc.qt.io/qt-6/qdialog.html#modal-prop
[2] https://docs.gtk.org/gtk3/property.Window.modal.html
[3] https://doc.qt.io/qt-6/qdialog.html#open

Change-Id: I099d990c614b592751e01791b48405c674b6ed3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184377
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>