sw a11y: No longer use unique_ptr for frame map
I can see no particular reason for using a
std::unique_ptr<SwAccessibleContextMap> for
SwAccessibleMap::mpFrameMap, instead of just a
SwAccessibleContextMap.

So far, an empty map was created if `mpFrameMap` was null
in both, SwAccessibleMap::GetContextImpl and
SwAccessibleMap::GetDocumentView_, and the member
was reset in either the dtor or in
SwAccessibleMap::RemoveContext when the map became empty.

I don't see any particular reason to have an explicit
null value (and check for it in multiple places) instead
of simply having an empty map by default.
Therefore, switch to using the latter, which simplifies
the code.

For most places, the previous null check can simply
be dropped as the overall logic remains the same
with an empty map (e.g. still nothing happens in case
of an empty map in most methods because no corresponding
entry is found in the empty map.), but keep an expclicit
check for these cases:

* At the beginning of SwAccessibleMap::BuildSelectedParas, replace
  the previous null check by checking for an empty map
  instead of just dropping the check, to avoid unnecessarily
  iterating over text frames.

* In the SwAccessibleMap dtor, also replace the
  previous null check by a check whether the
  map is empty.

Change-Id: I383b49fb84fc77ace80053355e9e69b144e6c72b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184441
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>