fix spellcheck dialog preventing replacement of "."
There appears to be special logic for replacing words that end
with a dot in the spellcheck dialog. Where the dot at the end is
ignored during replacement.

This special handling is only there for the spelldialog, i.e. it
is not there if you try to do the replacement via the context
menu or via IDocumentRedlineAccess.

e.g. If the text with the error text is only "." and the
replacement is "REPLACEMENT", clicking Correct button on the
dialog ends up with "REPLACEMENT."

Make it so that if the error text is just a dot, do not enforce
having a dot at the end of the replacement. Since it is obvious
the suggested replacement is for the dot.

It feels to me like the whole special handling for dot was
originally dictionary/suggestion error and was worked around in
the wrong place.

Change-Id: I65675ad18e5386f65ee978c0e92952bba29ae3cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184495
Tested-by: Jenkins
Reviewed-by: Sarper Akdemir <sarper.akdemir@allotropia.de>