sw a11y: Simplify paragraph parent retrieval in one place
The general way to retrieve the XAccessibleContext
from an XAccessible is to call XAccessible::getAccessibleContext,
not to query the XAccessible for the XAccessibleContext
interface.
(Doing the latter - as was done here - should still give the correct
result in practice at least, because SwAccessibleContext::getAccessibleParent
returns a reference to self.)

For SwAccessibleParagraph as a SwAccessibleContext subclass
however, there is SwAccessibleContext::getAccessibleParentImpl
(which also gets called by SwAccessibleContext::getAccessibleParent)
which returns a SwAccessibleContext right away, so use that
which is simpler and also removes the need to static_cast the
accessible to SwAccessibleContext altogether.

As a side not, using `pPara` for the var name of an accessible with
role TABLE_CELL also seemed a little misleading.

Change-Id: I38991674e806eb8f5ecf75327686651b6779c8ed
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184547
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins