Revert hiding of Start Center when loading modules or documents on macOS
In commit 331ea32ec181aff739eaebee328281b190cad003, the backing
frame's component was unset if there is no progress bar. This made
the Start Center's window empty so that loading large documents
wouldn't have the Start Center content displayed while the document
loads.

Unfortunately, this change has the side effect of making the Start
Center undisplayable if the LibreOffice recovery dialog is displayed
on launch and the "Discard All" button is pressed.

Change-Id: Iff571eb1751256f359cb0ef8a8eec7c094bcec68
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184554
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins