sw a11y: Don't use SwAccessibleChild when it's known to be a frame
SwAccessibleChild provides some abstraction from
whether the underlying object of interest is an
SwFrame, an SdrObject or a vcl::Window.

However, here in SwAccessibleMap::InvalidateCursorPosition,
the `SwAccessibleChild aFrameOrObj` is only ever used to
"wrap" SwFrame objects, and then SwAccessibleChild::GetSwFrame
gets called to get the underlying SwFrame from it again.

As far as SwAccessibleMap::InvalidateCursorPosition
is concerned, the only relevant thing that both,
the SwAccessibleChild ctor and
SwAccessibleChild::operator=(const SwFrame* pFrame)
effectively do is to assign SwAccessibleChild::mpFrame
that then gets returned in SwAccessibleChild::GetSwFrame.

The only other SwAccessibleChild method that gets called
is SwAccessibleChild::IsAccessible (for an OSL_ENSURE),
but that can now be replaced by call to the static
helper method SwAccessibleChild::IsFrameAccessible
introduced in previous commit

    Change-Id: Ic5c39cd7f7cb27a027adc6eeb7c77b1b6c4b5cba
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Apr 24 13:58:45 2025 +0200

        sw a11y: Extract static helper SwAccessibleChild::IsFrameAccessible

Avoid the extra level of indirection and simply use
a pointer to SwFrame directly instead of SwAccessibleChild.

This makes much more obvious that only the SwFrame* itself
is actually of interest.

Change-Id: I66cf2433f0d73ffab34e2a8a49f3903be77d6fdc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184581
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>