cool#11739 lok: sc: change dialog units per user
On lok different users might have different locales so they need
different field units in their dialogues. Right now the metric used is
always the one stored in the ScModule's AppOptions for all users.
With this change, for lok SfxModule::GetFieldUnit is used because it
already has the correct metric per lok user.

Change-Id: If2b047874a26deecc7c43936ab267b75eae101e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184567
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>