svx a11y: Don't duplicate frame selector a11y names as descriptions
The strings in RID_SVXSTR_FRMSEL_DESCRIPTIONS used for the
accessible description are exactly the same as those in
RID_SVXSTR_FRMSEL_TEXTS which is used for the accessible name.

That duplication doesn't provide any extra value (and might make
screen readers announce the same thing twice), so just return an
empty string for the accessible description instead.

Those are used e.g. for the borders in Writer's
"Format" -> "Paragraph" dialog, tab "Borders" for the
single borders in the drawing area labeled "User-defined".

Change-Id: I412ed482860fe1aecfc840b9c38491adf4f5b53e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184724
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>