tdf#166319 sw interdependent redlines: fix undo of del on top of ins
Load the document, it has an insertion and that has an additional
deletion inside the deletion. Reject the insert, undo, reject it again:
only the first half of the insertion gets rejected, instead of the
entire insert.

Reading the doc model xml dump, the problem is that the deletion is on
top of the insert (so the large insert gets split to 1) insert 2)
delete-on-insert and 3) insert), and the SwUndoRejectRedline only tracks
the delete for 2), there is no SwRedlineData in the action for the
previous insert part of 2).

Fix the problem by using SwUndoAcceptRedline in
DocumentRedlineManager::RejectRedlineRange(), which fixes the
inconsistency of commit 52fa7aed48632166e064e6a227e034f0981c4205
(tdf#157662 SW: redline: accept/reject done for all parts, 2023-10-14),
which used SwUndoRejectRedline + lcl_AcceptRedline() for the deletion
inside the insert.

This also means the custom depth is not needed for the created
SwUndoAcceptRedline, again making the undo action and the actual action
more consistent.

Change-Id: I7f38854efa01b08ab4fc705d7bd976eaa585c3b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184612
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>