tdf#165712 - sd: fix too small line spacing between text lines
in placeholders with shrink text on overflow

Allowing less Line spacing reduction in case of 0.925 scaling, which
would make our texts more readable when we are trying to shrink text
on overflow. Instead of decreasing the line spacing to 80%, should be
enough to decrease them to 90% and the font size should be decreased a bit more.
With that we are a bit more closer to mimics MSO text scaling.

Change-Id: Ie5a143cf7c0eeda30682850c59c1e6449a19234b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184704
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>