tdf#130857 qt weld: Don't multi-select tree items in single sel mode
Regardless of what selection mode is set for the tree view
displaying the model, QItemSelectionModel::select
adds to the selection and doesn't clear the previous one
unless the QItemSelectionModel::Clear flag is passed.

As a consequence, multiple tree view entries could be
selected after QtInstanceTreeView::select was called.

Prevent that by explicitly passing the
QItemSelectionModel::Clear flag if the tree view
is in single selection mode.

Prior to this change, multiple entries were e.g. selected
for the following scenario:

* start Writer
* "Tools" -> "AutoText", click "Categories" button
* in the dialog, add various new categories by
  typing text in the "Category" edit and clicking
  the "New" button

Behavior without this change in place for qt6 with
SAL_VCL_QT_USE_WELDED_WIDGETS=1: All newly
added categories would be selected at the same time.

Behavior with this change in place: Only the last selected
category is selected. This matches the behavior of other
VCL plugins.

Change-Id: I286b1e17b33d686553ae7c1533f07028b73bdfe5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184650
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins