tdf#60700 Unit tests for saving image icons and accelerators
For tdf#60700 we want to reduce the number of empty directories
under Configurations2/ that are created on every new document,
and we need tests to verify that the features that rely on those
directories continue to work.

My previous test at commit aaed16eef8b2d32ac02c8cb3a42ca25246cb0d33
covers most of the directories, which are directly created by
UIConfigurationManager. But Configurations2/images/ is created by
ImageManager and Configurations2/accelerator/ is created by
DocumentAcceleratorConfiguration, so we need separate tests for those
in order to exercise the codepaths.

This change adds those two tests.

Change-Id: Ie34b540f0c59de3bbb871b71b956d3387eb852c3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184646
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins