tdf#60700 stop creating empty UIConfiguration subdirs in ODFs
This change stops the creation of various subdirectories of the
Configurations2/ directory on new documents that do not have anything
that needs to be stored in those subdirectories. Previously, LibreOffice
would always create a bunch of empty subdirectories that were not
necessary.

Whenever an XStorage is opened readwrite on a path, it will immediately
create that path. LibreOffice was opening all of these paths readwrite.
So, the solution is to always open these paths as readonly, and then
when we need to write data to them we close them and re-open them as
readwrite. If we never need to write data to a path, then it won't be
created.

This patch specifically stops the creation of subdirectories that were
being created directly by UIConfigurationManager. There are other paths
that were being created by ImageManager and
DocumentAcceleratorConfiguration, and those paths will be handled by
other patches.

Test coverage was added in aaed16eef8b2d32ac02c8cb3a42ca25246cb0d33

$ CPPUNIT_TEST_NAME="Test testTdf69500" make CppunitTest_sw_odfexport

I also manually tested with the following procedure:

1. Open a new document in writer.
2. Tools > Customize > Toolbars.
3. Set scope to "Untitled 1".
4. Add a new toolbar "New Toolbar 1".
5. Add any command from Available Commands to the new toolbar.
6. OK.
7. Visually verify that the toolbar is present.
8. Save the document.
9. Verify that the toolbar is saved correctly inside the file.

$ unzip -l Untitled\ 1.odt
Archive:  Untitled 1.odt
  Length      Date    Time    Name
---------  ---------- -----   ----
       39  2025-04-25 23:35   mimetype
      353  2025-04-25 23:35   Configurations2/toolbar/custom_toolbar_2a773c3.xml
        0  2025-04-25 23:35   Configurations2/accelerator/
        0  2025-04-25 23:35   Configurations2/images/Bitmaps/
    12645  2025-04-25 23:35   styles.xml
      899  2025-04-25 23:35   manifest.rdf
     3563  2025-04-25 23:35   content.xml
     1012  2025-04-25 23:35   meta.xml
    14748  2025-04-25 23:35   settings.xml
      318  2025-04-25 23:35   Thumbnails/thumbnail.png
     1180  2025-04-25 23:35   META-INF/manifest.xml
---------                     -------
    34757                     11 files

10. Close and reopen the document.
11. Visually verify that the toolbar is present.

Change-Id: If0a93675204576b3b6f7fe9113cada398e4da2ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184655
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>