ScriptForge Don't watch close button of stored dialogs
Dialogs can either be stored in a library
or created from scratch.

When run in non-modal mode, the closing of
a stored dialog with the Close top-right cross
is done by LO correctly.

When the dialog is created from scratch,
a coded XTopWindowListener should do the job.

So far, the listener was activated for all dialogs.

Additionally the Console is a stored dialog.
When closed and relaunched, the processing
will now handle the window Close and
the built-in Close buttons identically, at
least apparently.

Change-Id: I29e0881989816fbf7e3e8bd12f59517bde4b3047
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184662
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins