tdf#130857 qt weld: Move TreeView method logic to overload for iters
weld::TreeView has a lot of methods that have two overloads
where one takes a row index and the other takes a weld::TreeIter.

In order to have to implement the logic in only one place,
QtInstanceTreeView so far implemented the logic in the
overload taking the row index, and the overload taking the
weld::TreeIter would forward to the overload taking the row index.

This was fine so far. As long the hierarchical structure of
"actual" trees wasn't supported and all instances of
QtInstanceTreeView where therefore just tables where
all tree view items are top level items (s.a. the assert
on `pParent` being null in QtInstanceTreeView::insert).
However, in order to support non-top-level items, i.e.
items that are children of other items, the approach
needs to change because the row-index based overload
only handles top-level items while a weld::TreeIter
can also represent non-toplevel items.

In order to prepare for that, switch the approach to
be the other way around: Move the logic from the overload
taking a row index to the overload taking a weld::TreeIter
and let the one taking a row index call the one taking
a weld::TreeIter instead.

Introduce new helper method QtInstanceTreeView::treeIter
to retrieve a TreeIter for a given (top-level) row index
and adjust/add other helper methods as needed now.

While at it, already start passing the QModelIndex::parent() [1]
as an additional param to some of the called methods
(instead of using the default, which would indicate
they are top-level items).
While this doesn't make any difference yet (as long
as all items are top-level items), it helps prepare
for supporting non-top-level items in the future.

No change of behavior in any of the already supported
dialogs is intended by this change.

[1] https://doc.qt.io/qt-6/qmodelindex.html#parent

Change-Id: Id829ce3a1052f0d8915f0b03a0b0786e2f3586dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184672
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins