tdf#130857 qt weld: Support non-top-level TreeView items
Drop the assert on `pParent` being null in
QtInstanceTreeView::insert and instead implement
handling for parent items.

In most of the cases, the corresponding QAbstractItemModel
methods have an optional (defaulting to `QModelIndex()`)
param to specify the parent. Pass that param instead
of using the default.

This builds on top of this previous commit whose
commit messages has some more details:

    Change-Id: Id829ce3a1052f0d8915f0b03a0b0786e2f3586dd
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Apr 26 17:00:01 2025 +0200

        tdf#130857 qt weld: Move TreeView method logic to overload for iters

Change-Id: Ibb1a119e3fc6f0fa878babbe526990df42d78336
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184673
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins