tdf#130857 qt weld: Keep col index in QtInstanceTreeView::copy_iterator
The model index retrieved via QtInstanceTreeView::modelIndex
may have an adjusted column number if the toggle button
feature is enabled.

However, if copying an iterator, the copy should
be equivalent to the source, so use the unmodified
QModelIndex from the source instead.

While the weld::TreeIter as used in LO doesn't have
a concept of columns but references a row
(and the column index is passed as an extra param
to methods where needed), not modifying the column
set in the QModelIndex can still be relevant as
QtInstanceTreeIter::equal directly compares the
model indices of the two iters, so the copy
wouldn't be considered equal if the column index
is different.

Change-Id: I3050daf17e7453cc62b1071647527ce888fd8a82
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184674
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>