tdf#130857 qt weld: Ensure tree view col count is same on all levels
Other than GtkTreeModel, where the same columns are always
used on all hierarchy levels, QAbstractItemModel allows
that to differ on each level/for each parent.

This was seen causing a crash in a WIP branch experimenting
with adding support for the "Tools" -> "AutoText" dialog
because

        const QModelIndex aIndex = modelIndex(nPos, 0, aParentIndex);
        QStandardItem* pItem = itemFromIndex(aIndex);
        if (pStr)
            pItem->setText(toQString(*pStr));

inside QtInstanceTreeView::insert would result in
trying to retrieve the index/item for column 0
when there were actually no columns in the model
for the non-top-level item.

Make sure the amount of columns matches by inserting
them when there are none yet.

Change-Id: I3bca4eb2e036e682298db7a205279bfddf0e8805
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184675
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins