tdf#166258 Refactor throttling of Skia/Metal flushes
Throttling of Skia/Metal flushes is needed for tdf#163945 but the
throttling logic was being applied to too frequently. So throttle
based on the time interval of the last Skia/Metal flush and if the
flush is too soon after the last flush (i.e. 1/200th of a second),
schedule the flush in the Skia timer.

Change-Id: Ib4e526ec26ec8b8d36bf3a9776bfe49d153d759f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184681
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>