Related: tdf#165448 move parent window to front after closing window
When a floating window such as the dropdown list in the
font combobox is open when selecting any of the menu items
inserted by macOS in the windows menu, the parent window
will be hidden. So if there is a key window, force the key
window back to the front.

Previously, delaying closing of the window was used to avoid
this bug, but that caused Skia/Metal to crash when the
window was released before the delayed close occurred. This
crash was found when rapidly dragging the border between two
column headings side to side in a Calc document.

Change-Id: I8511fe73e59c49f57d9179f5308e4b7a393809b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184690
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>