oox: Don't substitute fonts if embedding fonts is enabled
This is needed to prevent that we embed a font but then export
a different (substituted) font name into the document, which
would defeat the purpose of embedding in the first place.

This is not ideal either as a we probably need better control of
what we can just substitue and what we need to embed.

Change-Id: Icd29789ceb739b6950576379358a96cc0449b3b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184696
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 0c1b3c1c97bbfdc8d8f84b36fb5879400095c5e0)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184824
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>