prefer the last fontFamily name, like fontconfig
There was a report that font names of the "Helvetica Neue LT Pro"
font family were different, when the fonts were added to the system,
and when the fonts were added via AddTempDevFont() (remote font config
in COOL).
E.g. the font names were expected to be
- HelveticaNeueLT Pro 45 Lt
- HelveticaNeueLT Pro 65 Md
- etc.
Instead, we always got "Helvetica Neue LT Pro" which did not match, and
we had a single font name for multiple fonts.
In these fonts, in the name table there are two platform specific tables,
macintosh and windows. The fontFamily values are different in these
tables. In vcl/unx/generic/fontmanager/fontconfig.cxx we reverse the
order of names. In the case of these fonts the windows table is used.
With this patch we do the same for the AddTempDevFont() route.

Change-Id: Iaceaabb406dbe63bc76f1066169268de6364b8e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184698
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
(cherry picked from commit 36de29fe80bfc451d026b986c331afe7ae1129be)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184797
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Tested-by: Andras Timar <andras.timar@collabora.com>