lok: use unique ids in sidebar
Panel was shared id in the past (we needed to be able to trigger
refresh for initial mobile-wizard where no welding was done).
So we needed the same id to catch any instance and do a full update
of a sidebar.

Now we have welding so we can use single widget as target
and avoid sending full JSON on context change. Show/Hide
panel functionality should be used now and thanks to that
save some bandwidth.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: Ib564b608cb4de04185fb89832ad00bea9bcc10bd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184717
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>