svx a11y: Report checked/checkable state for borders
AccFrameSelectorChild is using an accessible role of
AccessibleRole::CHECK_BOX and when the status/visibility
of a border changes in FrameSelectorImpl::SetBorderState,
an a11y STATE_CHANGED event was already sent for the
CHECKED state. However, the CHECKABLED and CHECKED states
themselves were never reported in
AccFrameSelectorChild::getAccessibleStateSet.

Extend AccFrameSelectorChild::getAccessibleStateSet
accordingly to match the criterion used in
FrameSelectorImpl::SetBorderState.

The state now shows up as expected with the qt6 VCL plugin
in Accerciser for the border's a11y objects in Writer's
"Format" -> "Paragraph" dialog, tab "Borders" when
one of the borders in the drawing area under the "User-defined"
label gets selected in Accerciser's tree view of the
a11y hierarchy.

In addition, Orca now announces the state as well, e.g.
"tick box ticked" when such a border receives focus,
and toggling the state using the space key results in the
new state getting announced ("not ticked"/"ticked" when
using Orca with en_GB locale).

Change-Id: Ia6dbe668a32829a94bc31726a434cad4ac1159cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/184726
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>